requirements.txt
faiss-cpu
langchain-groq
PyPDF2
langchain
streamlit
langchain_community
python-dotenv
pypdf
pytesseract
pillow
tiktoken
pandas
unstructured
python-docx
reportlab
PyMuPDF
langchain_groq
sentence-transformers
langchain_huggingface
uvicorn[standard] 
fastapi
python-multipart
markdown
langchain
bibtexparser
crewai